requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
pydantic[email]==2.10.3
slowapi==0.1.9
pytest==8.3.4
httpx==0.28.1
